refactor: move bench gate types + evaluation to homeboy-extension-contract
Continues the bench cluster breakdown. The three bench gate types (BenchGate,
BenchGateOp, BenchGateResult) plus their evaluation impls (BenchGate::evaluate/
evaluate_actual, BenchGateOp::as_str) move to the contract crate.

Key finding: although bench/gate.rs imports core's HomeboyGate* types, those are
used ONLY by the evaluate_gates/normalized_gate_results FUNCTIONS (which also
consume BenchResults + budget_findings) - not by the gate type definitions or
their impls. The impls are pure: they only read BenchMetrics (already in
contract), match on BenchGateOp, and format strings. So the types + impls move
cleanly; the two core-coupled functions stay in core.

Widened BenchGate::evaluate from pub(crate) to pub for cross-crate access. Core
re-exports the types (pub use) so bench module paths stay stable.

BenchGate/BenchGateResult are field dependencies of BenchScenario - another
blocker cleared toward the top-level bench result cluster.
