+++ crates/homeboy-extension-contract/src/bench_gate.rs
@@ -1,0 +1,85 @@
+//! Pure bench gate contract types + their evaluation logic.
+
+use serde::{Deserialize, Serialize};
+
+use crate::BenchMetrics;
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct BenchGate {
+    pub metric: String,
+    pub op: BenchGateOp,
+    pub value: f64,
+}
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "lowercase")]
+pub enum BenchGateOp {
+    Eq,
+    Gte,
+    Lte,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+#[serde(deny_unknown_fields)]
+pub struct BenchGateResult {
+    pub metric: String,
+    pub op: BenchGateOp,
+    pub expected: f64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub actual: Option<f64>,
+    pub passed: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub reason: Option<String>,
+}
+
+impl BenchGate {
+    pub fn evaluate(&self, scenario_id: &str, metrics: &BenchMetrics) -> BenchGateResult {
+        let actual = metrics.get(&self.metric);
+        self.evaluate_actual(&format!("scenario `{}`", scenario_id), actual)
+    }
+
+    pub fn evaluate_actual(&self, scope: &str, actual: Option<f64>) -> BenchGateResult {
+        let passed = actual
+            .map(|value| match self.op {
+                BenchGateOp::Eq => value == self.value,
+                BenchGateOp::Gte => value >= self.value,
+                BenchGateOp::Lte => value <= self.value,
+            })
+            .unwrap_or(false);
+        let reason = if passed {
+            None
+        } else {
+            Some(match actual {
+                Some(value) => format!(
+                    "{} gate failed: {} {} {} (actual {})",
+                    scope,
+                    self.metric,
+                    self.op.as_str(),
+                    self.value,
+                    value
+                ),
+                None => format!("{} gate failed: metric `{}` is missing", scope, self.metric),
+            })
+        };
+
+        BenchGateResult {
+            metric: self.metric.clone(),
+            op: self.op,
+            expected: self.value,
+            actual,
+            passed,
+            reason,
+        }
+    }
+}
+
+impl BenchGateOp {
+    pub fn as_str(self) -> &'static str {
+        match self {
+            BenchGateOp::Eq => "eq",
+            BenchGateOp::Gte => "gte",
+            BenchGateOp::Lte => "lte",
+        }
+    }
+}
